Reduce login hops
